+++ externalstorage/src/main/scala/fluence/swarm/crypto/Secp256k1Signer.scala
@@ -1,20 +1,4 @@
-/*
- * Copyright 2018 Fluence Labs Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package fluence.swarm
+package fluence.swarm.crypto
 import fluence.crypto.{Crypto, CryptoError}
 import org.web3j.crypto.{ECKeyPair, Sign}
 import scodec.bits.ByteVector
@@ -23,6 +7,7 @@ import scala.util.Try
 
 /**
  * Default signing logic adapted from the BitcoinJ ECKey.
+ *
  * @see https://github.com/bitcoinj/bitcoinj/blob/master/core/src/main/java/org/bitcoinj/core/ECKey.java
  */
 private[swarm] object Secp256k1Signer {
