+++ externalstorage/src/main/scala/fluence/swarm/requests/InitializeMutableResourceRequest.scala
@@ -19,8 +19,9 @@ package fluence.swarm.requests
 import cats.Monad
 import cats.data.EitherT
 import fluence.crypto.Crypto.Hasher
-import fluence.swarm.Secp256k1Signer.Signer
+import fluence.swarm.crypto.Secp256k1Signer.Signer
 import fluence.swarm._
+import fluence.swarm.meta.{MetaHash, RootAddr, Signature}
 import io.circe.Encoder
 import io.circe.generic.semiauto._
 import scodec.bits.ByteVector
@@ -64,7 +65,7 @@ object InitializeMutableResourceRequest {
   import MetaHash._
   import RootAddr._
   import Signature._
-  import ByteVectorJsonCodec._
+  import fluence.swarm.helpers.ByteVectorJsonCodec._
 
   implicit val initializeRequestEncoder: Encoder[InitializeMutableResourceRequest] = deriveEncoder
